requirements.txt
beautifulsoup4
flask
google-auth
google-auth-oauthlib
google-api-python-client
psycopg2-binary
cloudinary
gunicorn
